+++ api/build.gradle
@@ -25,8 +25,7 @@ dependencies {
     implementation 'org.skyscreamer:jsonassert:1.5.0'
     implementation 'com.fasterxml.jackson.dataformat:jackson-dataformats-text:2.13.3'
     implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.13.3'
-  implementation 'commons-io:commons-io:2.11.0'
-  compileOnly 'org.projectlombok:lombok'
+    implementation 'commons-io:commons-io:2.11.0'
     runtimeOnly 'org.postgresql:postgresql'
     runtimeOnly 'com.h2database:h2'
     annotationProcessor 'org.projectlombok:lombok'
