--- api/src/main/java/com/mtech/vmcs/service/impl/initialfiledata/IInitialFileDataAdapter.java
@@ -1,10 +1,0 @@
-package com.mtech.vmcs.service.impl.initialfiledata;
-
-
-import java.io.IOException;
-
-public interface IInitialFileDataAdapter {
-  InitialFileDataBean read() throws IOException;
-
-  void write(InitialFileDataBean initialFileData) throws IOException;
-}
