+++ api/src/main/java/com/mtech/vmcs/service/impl/initialfiledata/InitialJsonFileDataAdapter.java
@@ -1,6 +1,8 @@
 package com.mtech.vmcs.service.impl.initialfiledata;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mtech.vmcs.model.entity.InitialFileData;
+import com.mtech.vmcs.service.IInitialFileDataAdapter;
 
 import java.io.File;
 import java.io.IOException;
@@ -13,13 +15,13 @@ public class InitialJsonFileDataAdapter implements IInitialFileDataAdapter {
   }
 
   @Override
-  public InitialFileDataBean read() throws IOException {
+  public InitialFileData read() throws IOException {
     ObjectMapper objectMapper = new ObjectMapper();
-    return objectMapper.readValue(this.file, InitialFileDataBean.class);
+    return objectMapper.readValue(this.file, InitialFileData.class);
   }
 
   @Override
-  public void write(InitialFileDataBean initialFileData) throws IOException {
+  public void write(InitialFileData initialFileData) throws IOException {
     ObjectMapper objectMapper = new ObjectMapper();
     objectMapper.writeValue(this.file, initialFileData);
   }
