+++ api/src/main/java/com/mtech/vmcs/service/impl/initialfiledata/InitialYamlFileDataAdapter.java
@@ -2,6 +2,8 @@ package com.mtech.vmcs.service.impl.initialfiledata;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.mtech.vmcs.model.entity.InitialFileData;
+import com.mtech.vmcs.service.IInitialFileDataAdapter;
 
 import java.io.File;
 import java.io.IOException;
@@ -14,13 +16,13 @@ public class InitialYamlFileDataAdapter implements IInitialFileDataAdapter {
   }
 
   @Override
-  public InitialFileDataBean read() throws IOException {
+  public InitialFileData read() throws IOException {
     ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
-    return objectMapper.readValue(this.file, InitialFileDataBean.class);
+    return objectMapper.readValue(this.file, InitialFileData.class);
   }
 
   @Override
-  public void write(InitialFileDataBean initialFileData) throws IOException {
+  public void write(InitialFileData initialFileData) throws IOException {
     ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
     objectMapper.writeValue(this.file, initialFileData);
   }
